refactor: remove SuggestionUseCase and integrate completion suggestions directly
