+++ domain/src/commonMain/kotlin/io/github/arashiyama11/dncl_ide/domain/DomainModule.kt
@@ -4,7 +4,6 @@ import io.github.arashiyama11.dncl_ide.domain.usecase.FileNameValidationUseCase
 import io.github.arashiyama11.dncl_ide.domain.usecase.FileUseCase
 import io.github.arashiyama11.dncl_ide.domain.usecase.NotebookFileUseCase
 import io.github.arashiyama11.dncl_ide.domain.usecase.SettingsUseCase
-import io.github.arashiyama11.dncl_ide.domain.usecase.SuggestionUseCase
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
@@ -12,6 +11,5 @@ val domainModule = module {
     singleOf(::FileNameValidationUseCase)
     singleOf(::FileUseCase)
     singleOf(::SettingsUseCase)
-    singleOf(::SuggestionUseCase)
     singleOf(::NotebookFileUseCase)
 }
